remove old main.tsx and app files
