+++ index.html
@@ -10,8 +10,7 @@
 
 <body>
   <div id="root"></div>
-  <script type="module" src="/src/tanstack-main.tsx"></script>
-  <!-- <script type="module" src="/src/main.tsx"></script> -->
+  <script type="module" src="/src/main.tsx"></script>
 </body>
 
 </html>
